[DDW-821] Renames status domain to nodes, fixes paths on all import statements within nodes dir
