+++ tests/nodes/e2e/steps/about-dialog.js
