+++ tests/nodes/e2e/steps/app-steps.js
@@ -1,9 +1,9 @@
 // @flow
 import { Given, When, Then } from 'cucumber';
 import { expect } from 'chai';
-import { waitUntilTextInSelector } from '../helpers/shared-helpers';
-import { refreshClient, waitForCardanoNodeToExit, waitForDaedalusToExit } from '../helpers';
-import type { Daedalus } from '../../types';
+import { waitUntilTextInSelector } from '../../../common/e2e/steps/helpers';
+import { refreshClient, waitForCardanoNodeToExit, waitForDaedalusToExit } from './helpers';
+import type { Daedalus } from '../../../types';
 
 declare var daedalus: Daedalus;
