+++ tests/nodes/e2e/steps/inject-fault.js
@@ -1,4 +1,8 @@
+// @flow
 import { When } from 'cucumber';
+import type { Daedalus } from '../../../types';
+
+declare var daedalus: Daedalus;
 
 When(/^I inject fault named "([^"]*)"$/, async function(faultName) {
   await this.client.executeAsync((name, done) => {
