+++ tests/nodes/e2e/steps/trouble-syncing-notification-steps.js
@@ -1,5 +1,9 @@
+// @flow
 import { When, Then } from 'cucumber';
-import { waitUntilTextInSelector } from '../helpers/shared-helpers';
+import { waitUntilTextInSelector } from '../../../common/e2e/steps/helpers';
+import type { Daedalus } from '../../../types';
+
+declare var daedalus: Daedalus;
 
 When(
   /^I arbitrarily set the local block height to half the network block height$/,
